Add 'runtest-junitxml.py' to generate junit-xml output
This is an alternative executor to runm3uatest.  It generates output in
junit-xml format, which is very useful for integrating reports form
automatic testsuite executing by C/I systems such as jenkins.

It's a bit sad to introduce yet another programming language beyond
lisp, shell and C - but then, python has a ready-to use junit XML module
and contrary to scheme/lisb, I know how to write it...